Remove fixtures in favor of generators

+++ test/util.js
@@ -1,11 +1,14 @@
+const _ = require('lodash');
 const mongoose = require('mongoose');
+const generators = require('./generators');
 
 module.exports = {
-  async seedDatabase(fixtures = {}) {
-    await mongoose.connection.dropDatabase();
+  async seedDatabase(config = {}) {
+    if(!config.keepDatabase) await mongoose.connection.dropDatabase();
     await mongoose.model('User').addUser({username: 'admin', password: 'admin'});
-    for(let key of Object.keys(fixtures)) {
-      await mongoose.model(key).insertMany(fixtures[key]);
+    for(let model of config.models || []) {
+      const generator = config.generator || generators[model.key];
+      await mongoose.model(model.key).insertMany(_.times(model.count || 10, generator));
     }
   },
   async getAuthToken(username = 'admin') {
